Reorganized files in folders in assets
