+++ src/components/Cards/ProductCard.jsx
@@ -1,9 +1,9 @@
 import { IoCartSharp } from "react-icons/io5";
 
-export default function Card({ image, name, children, url, ...props }) {
+export default function ProductCard({ image, name, children, url, ...props }) {
   const { available = true, mobile = false } = props;
   return (
-    <article className="rounded-md overflow-hidden relative max-w-sm group shadow-md mx-auto transform hover:-translate-y-1 duration-300 hover:shadow-xl cursor-pointer">
+    <article className="rounded-md overflow-hidden relative max-w-xs group shadow-md mx-auto transform hover:-translate-y-1 duration-300 hover:shadow-xl cursor-pointer">
       <div className="overflow-hidden">
         <img
           className="w-full h-auto transform hover:scale-110 duration-200"
@@ -11,9 +11,9 @@ export default function Card({ image, name, children, url, ...props }) {
           alt={name}
         />
       </div>
-      <div className="p-7 my-auto pb-12 bg-cdln-blue-50">
-        <h1 className="text-4xl font-semibold text-gray-800 mt-4">{name}</h1>
-        <p className="text-lg md:text-2xl text-gray-400 mt-4 leading-relaxed">
+      <div className="p-7 my-auto pb-5 bg-cdln-blue-50">
+        <h1 className="text-3xl font-semibold text-gray-800 mt-4">{name}</h1>
+        <p className="text-sm md:text-2xl text-gray-400 mt-4 leading-relaxed">
           {children}
         </p>
       </div>
